refactor: change core package
- core 패키지명 변경
