+++ see-realview/src/main/java/com/see/realview/_core/exception/CustomException.java
@@ -1,6 +1,6 @@
-package com.see.realview.core.exception;
+package com.see.realview._core.exception;
 
-import com.see.realview.core.response.ResponseData;
+import com.see.realview._core.response.ResponseData;
 import org.springframework.http.HttpStatus;
 
 public interface CustomException {
