+++ see-realview/src/main/java/com/see/realview/_core/exception/client/NotFoundException.java
@@ -1,14 +1,17 @@
-package com.see.realview.core.exception;
+package com.see.realview._core.exception.client;
 
-import com.see.realview.core.response.ErrorData;
-import com.see.realview.core.response.ResponseData;
+import com.see.realview._core.exception.CustomException;
+import com.see.realview._core.exception.ExceptionStatus;
+import com.see.realview._core.response.ErrorData;
+import com.see.realview._core.response.ResponseData;
 import org.springframework.http.HttpStatus;
 
-public class UnauthorizedException extends RuntimeException implements CustomException {
+public class NotFoundException extends RuntimeException implements CustomException {
+
     private final ExceptionStatus exception;
 
 
-    public UnauthorizedException(ExceptionStatus exception) {
+    public NotFoundException(ExceptionStatus exception) {
         this.exception = exception;
     }
