+++ see-realview/src/main/java/com/see/realview/_core/exception/server/ServerException.java
@@ -1,15 +1,17 @@
-package com.see.realview.core.exception;
+package com.see.realview._core.exception.server;
 
-import com.see.realview.core.response.ErrorData;
-import com.see.realview.core.response.ResponseData;
+import com.see.realview._core.exception.CustomException;
+import com.see.realview._core.exception.ExceptionStatus;
+import com.see.realview._core.response.ErrorData;
+import com.see.realview._core.response.ResponseData;
 import org.springframework.http.HttpStatus;
 
-public class BadRequestException extends RuntimeException implements CustomException {
+public class ServerException extends RuntimeException implements CustomException {
 
     private final ExceptionStatus exception;
 
 
-    public BadRequestException(ExceptionStatus exception) {
+    public ServerException(ExceptionStatus exception) {
         this.exception = exception;
     }
